backend/requirements.txt
fastapi
uvicorn
tensorflow
pillow
numpy
tensorflow-hub
